+++ config/default.js
@@ -239,11 +239,18 @@ module.exports = config.createConfig({dirs: [__dirname, 'client']}, {
 
 		/**
 		 * A name of the component to build demo examples or tests
+		 *
+		 * @cli demo-page
+		 * @env DEMO_PAGE
+		 *
+		 * @returns {string}
 		 */
-		demoPage: o('demo-page', {
-			env: true,
-			default: 'p-v4-components-demo'
-		}),
+		demoPage() {
+			return o('demo-page', {
+				env: true,
+				default: 'p-v4-components-demo'
+			});
+		},
 
 		/**
 		 * Port for a test server
@@ -382,10 +389,11 @@ module.exports = config.createConfig({dirs: [__dirname, 'client']}, {
 
 			if (enabled) {
 				return {
-					type: this.config.build.ci ? 'println' : 'progressbar',
 					opts: {
 						clearOnComplete: true,
 						stopOnComplete: true,
+						forceRedraw: true,
+						noTTYOutput: this.config.build.ci,
 						hideCursor: null
 					}
 				};
@@ -1044,7 +1052,8 @@ module.exports = config.createConfig({dirs: [__dirname, 'client']}, {
 			display: 'standalone',
 			orientation: 'portrait',
 			version: 1.0,
-			logging: false
+			logging: false,
+			manifestName: 'manifest.json'
 		};
 	},