+++ src/core/component/directives/aria/roles/index.ts
@@ -1,0 +1,20 @@
+/*!
+ * V4Fire Client Core
+ * https://github.com/V4Fire/Client
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Client/blob/master/LICENSE
+ */
+
+export * from 'core/component/directives/aria/roles/dialog';
+export * from 'core/component/directives/aria/roles/tablist';
+export * from 'core/component/directives/aria/roles/tab';
+export * from 'core/component/directives/aria/roles/tabpanel';
+export * from 'core/component/directives/aria/roles/controls';
+export * from 'core/component/directives/aria/roles/combobox';
+export * from 'core/component/directives/aria/roles/listbox';
+export * from 'core/component/directives/aria/roles/option';
+export * from 'core/component/directives/aria/roles/tree';
+export * from 'core/component/directives/aria/roles/treeitem';
+
+export { AriaRoleEngine, EngineOptions } from 'core/component/directives/aria/roles/interface';