+++ src/core/component/directives/aria/roles/option/interface.ts
@@ -1,0 +1,14 @@
+/*!
+ * V4Fire Client Core
+ * https://github.com/V4Fire/Client
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Client/blob/master/LICENSE
+ */
+
+import type { HandlerAttachment } from 'core/component/directives/aria/roles/interface';
+
+export class OptionParams {
+	isSelected: boolean = false;
+	'@change': HandlerAttachment = () => undefined;
+}