+++ src/core/component/directives/aria/roles/tab/interface.ts
@@ -1,0 +1,17 @@
+/*!
+ * V4Fire Client Core
+ * https://github.com/V4Fire/Client
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Client/blob/master/LICENSE
+ */
+
+import type { HandlerAttachment } from 'core/component/directives/aria/roles/interface';
+
+export class TabParams {
+	isFirst: boolean = false;
+	isSelected: boolean = false;
+	hasDefaultSelectedTabs: boolean = false;
+	orientation: string = 'false';
+	'@change': HandlerAttachment = () => undefined;
+}